feat : Lotto MyPage API - 이전 추천 받았던 모든 번호 가져오기 - 이번주, 추천 받은 번호가 있으면 '로그인 추천' 시 해당 번호 노출   (추천 알고리즘 추가 예정)

+++ Lotto_springboot_dev/src/main/java/com/lotto/dev/controller/GameController.java
@@ -23,7 +23,7 @@ public class GameController {
     }
 
     @GetMapping("/round/{id}")
-    public ResponseEntity<GameResponseDto> getBall(@PathVariable long id) {
+    public ResponseEntity<GameResponseDto> getBall(@PathVariable int id) {
         return ResponseEntity.ok(gameService.getBall(id));
     }
 }
